Hash.new takes capacity:, and Integer#size, #pow, #rationalize agree with MRI
Hash.new accepts the capacity: keyword (3.4) and refuses any other, with or
without a default value or block; a Hash passed as a plain argument is still
the default value, and a second plain argument is an arity error.
Hash#transform_keys/values and their bang forms return sized enumerators.
Integer#size is 8 for a fixnum and the byte length of a bignum's magnitude;
Integer#pow with an explicit nil modulus is a TypeError; Integer#rationalize
takes (and ignores) a tolerance.

core 422 -> 412 (parallel sweep), library 129, language 110, IronRuby.Tests at reference.

+++ Src/Libraries/Builtins/FixnumOps.cs
@@ -61,7 +61,8 @@ namespace IronRuby.Builtins {
 
         [RubyMethod("size")]
         public static int Size(int self) {
-            return sizeof(int);
+            // the size of MRI's fixnum on a 64-bit machine: a C long
+            return sizeof(long);
         }
     }
 }
